requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.0.6
django-crispy-forms==1.14.0
gunicorn==20.1.0
numpy==1.23.2
pandas==1.4.4
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
XlsxWriter==3.0.3
openpyxl==3.0.10